requirements.txt
altair>=4.2.1,<5.0.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
ddgs==9.13.0
duckdb==1.5.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
playwright==1.58.0
primp==1.2.2
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.1
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
urllib3==2.6.3
playwright-stealth
great-expectations==0.18.22
pdfplumber
python-docx
prefect
